В команды добавлена проверка login и password

+++ src/main/java/ru/lab6/server/model/command/ClearCommand.java
@@ -1,6 +1,8 @@
 package ru.lab6.server.model.command;
 
+import ru.lab6.common.parameters.CreationParameters;
 import ru.lab6.common.parameters.EmptyParameters;
+import ru.lab6.common.parameters.LoginParameters;
 import ru.lab6.common.parameters.Parameters;
 import ru.lab6.common.response.Response;
 import ru.lab6.server.model.ApplicationContext;
@@ -18,6 +20,18 @@ public class ClearCommand implements Command {
             applicationContext.getUserDao().deleteAll();
             return new Response().setErrorResponse("ошибка параметров команды", "");
         } else {
+
+            EmptyParameters emptyParameters = (EmptyParameters) parameters;
+
+            LoginParameters loginParameters = new LoginParameters();
+            loginParameters.login = emptyParameters.login;
+            loginParameters.password = emptyParameters.password;
+            Response response = applicationContext.getCommands().get("login").execute(loginParameters);
+
+            if (!response.getStatus().equals("ok")) {
+                return response;
+            }
+
             applicationContext.getRepository().deleteAll();
             return new Response().setEmptyResult();
         }
